Cas INT_MIN gere sur conversion integer

+++ libft-srcs/ft_putnbr_fd.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putnbr_fd.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lchapren <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: lchapren <lchapren@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/07 15:30:29 by lchapren          #+#    #+#             */
-/*   Updated: 2019/11/07 17:52:23 by lchapren         ###   ########.fr       */
+/*   Updated: 2019/12/18 11:39:53 by lchapren         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putnbr_fd(int n, int fd)
+void	ft_putnbr_fd(long int n, int fd)
 {
 	long int nbl;
